Add files via upload
This is the data 1928-2024 for annual realized volatility for Standard & Poor 500 and its predecessor, Standard & Poor 90, based on daily log price returns, nominal (not inflation-adjusted). Also, the code for fitting an autoregressive model of order 1 on the log scale. The data matches the one on my web page 'Annual 1928-2024, my undergraduate students'. The code corresponds to the blog https://my-finance.org/2025/02/19/updated-annual-volatility-for-2024/ and https://my-finance.org/2025/01/22/annual-volatility/